+++ client/components/CodeText.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import TextField from '@mui/material/TextField';
 import { useSelector, useDispatch } from 'react-redux'
 
-import { userEditText } from '../reducers/reducer';
+import { userEditText } from '../redux/reducer';
 
 // This component will render the code received from the fetch request to the server
 // Also, in the stretch feature, as the user edits the code, it will be saved to the database
