+++ client/containers/ButtonContainer.jsx
@@ -5,7 +5,7 @@ import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 import DoneAllIcon from '@mui/icons-material/DoneAll';
 import { useSelector, useDispatch } from 'react-redux'
 
-import { copyText, changeIcon } from '../reducers/reducer';
+import { copyText, changeIcon } from '../redux/reducer';
 
 // This container wraps:
 // 2) the button that copies the code to the clipboard
